Move db_cache into mod.rs
The linter was complaining about module inception otherwise.

https://rust-lang.github.io/rust-clippy/master/index.html#module_inception
